2019_10_22 | django_mtom

+++ 03_django/03_django_form/articles/views.py
@@ -1,3 +1,4 @@
+import hashlib
 from IPython import embed
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
@@ -17,7 +18,7 @@ def index(request):
     request.session.modified = True
 
     articles = Article.objects.all()
-    context = {'articles': articles,'visits_num': visits_num}
+    context = {'articles': articles,'visits_num': visits_num,}
     return render(request, 'articles/index.html', context)
 
 @login_required
@@ -28,7 +29,9 @@ def create(request):
         form = ArticleForm(request.POST)
         # form 이 유효한지 체크한다.
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save() 
             return redirect(article)
     else:
         form = ArticleForm()
@@ -51,20 +54,26 @@ def detail(request, article_pk):
 def delete(request, article_pk):
     if request.user.is_authenticated:
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
+        if request.user == article.user:
+            article.delete()
+        else:
+            return redirect(article)
     return redirect('articles:index')
 
 
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            article = form.save()
-            return redirect(article)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                article = form.save()
+                return redirect(article)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:index')
     context = {'form': form, 'article': article,}
     return render(request, 'articles/form.html', context)
 
@@ -74,9 +83,9 @@ def comments_create(request, article_pk):
     if request.user.is_authenticated:
         comment_form = CommentForm(request.POST)
         if comment_form.is_valid():
-            # 객체를 Create 하지만, db 에 레코드는 작성하지 않는다.
             comment = comment_form.save(commit=False)
             comment.article_id = article_pk
+            comment.user = request.user
             comment.save()
     return redirect('articles:detail', article_pk)
 
@@ -85,6 +94,24 @@ def comments_create(request, article_pk):
 def comments_delete(request, article_pk, comment_pk):
     if request.user.is_authenticated:
         comment = get_object_or_404(Comment, pk=comment_pk)
-        comment.delete()
+        if comment.user == request.user:
+            comment.delete()
         return redirect('articles:detail', article_pk)
     return HttpResponse('You ar Unauthorized', status=401)
+
+
+def like(request, article_pk):
+    article = get_object_or_404(Article, pk=article_pk)
+
+    if article.like_users.filter(pk=request.user.pk).exists():
+        article.like_users.remove(request.user)
+    else:
+        article.like_users.add(request.user)
+    return redirect('articles:index')
+
+    # # 해당 게시글에 좋아요를 누른 사람들 중에서 현재 접속유저가 있다면 좋아요를 취소
+    # if request.user in article.like_users.all():
+    #     article.like_users.remove(request.user) # 좋아요 취소
+    # else:
+    #     article.like_users.add(request.user) # 좋아요
+    # return redirect('articles:index')
